Extract function out from component
